backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
python-dotenv==0.19.0
sqlalchemy==1.4.23
pydantic==1.8.2
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
openai-whisper==20231117
transformers==4.30.2
torch==2.0.1
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
nltk==3.8.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
pytest==7.3.1
httpx==0.24.0
alembic==1.7.7 